// src/app/_admin/Models/userViewInfoModel.ts
import { RoundDateModel } from './roundDatesModel';

export class UserViewInfoModel {
    // firstRound: RoundDateModel;
    // secondRound: RoundDateModel;
    firstRoundStartDate: Date;
    firstRoundEndDate: Date;
    secondRoundStartDate: Date;
    secondRoundEndDate: Date;
    policyURL: string;
    policyFilePath: string;
}